Reorganize the hack directory.
