+++ .github/dependabot.yml
@@ -5,7 +5,7 @@ updates:
   - package-ecosystem: "gomod"
     directories:
       - "/"
-      - "/hack/internal/tools"
+      - "/hack/tools"
       - "/site"
     schedule:
       interval: "weekly"
@@ -40,11 +40,12 @@ updates:
   - package-ecosystem: "docker"
     directories:
       - "/cmd/kueueviz/frontend"
-      - "/hack/agnhost"
-      - "/hack/cypress"
-      - "/hack/depcheck"
-      - "/hack/shellcheck"
-      - "/hack/multikueue/secretreader"
+      - "/hack/testing/agnhost"
+      - "/hack/testing/cypress"
+      - "/hack/testing/depcheck"
+      - "/hack/testing/ray"
+      - "/hack/testing/secretreader"
+      - "/hack/testing/shellcheck"
     schedule:
       interval: "weekly"
     labels:
