+++ hack/testing/OWNERS
@@ -1,0 +1,10 @@
+# See the OWNERS docs at https://go.k8s.io/owners
+
+filters:
+  ".*":
+    approvers:
+      - test-approvers
+
+  "Dockerfile$":
+    approvers:
+      - dependency-approvers
