+++ hack/testing/compare-performance.sh
