+++ hack/testing/e2e-test.sh
@@ -19,9 +19,9 @@ set -o nounset
 set -o pipefail
 
 SOURCE_DIR="$(cd "$(dirname -- "${BASH_SOURCE[0]}")" && pwd -P)"
-ROOT_DIR="$SOURCE_DIR/.."
+ROOT_DIR="$SOURCE_DIR/../.."
 
-# shellcheck source=hack/e2e-common.sh
+# shellcheck source=hack/testing/e2e-common.sh
 source "${SOURCE_DIR}/e2e-common.sh"
 
 function cleanup {
