+++ hack/testing/get-build-logs.sh
