+++ hack/testing/shellcheck/verify.sh
