+++ hack/tools/yaml-processor/yamlproc/txtinserter.go
